requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
python-dotenv
torch
torchvision
numpy
pillow
opencv-python
